Modified AlertBanne,AnimatedSection,ArticileHeader
