+++ examples/kit-nextjs-article-starter/src/components/alert-banner/AlertBanner.dev.tsx
@@ -1,12 +1,40 @@
 import { useState } from 'react';
-import { Text } from '@sitecore-content-sdk/nextjs';
 import { ButtonBase } from '@/components/button-component/ButtonComponent';
 import { Button } from '@/components/ui/button';
 import { X } from 'lucide-react';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { cn } from '@/lib/utils';
 import { NoDataFallback } from '@/utils/NoDataFallback';
-import { AlertBannerProps } from './alert-banner.props';
+import { Field, ImageField, LinkField, Text } from '@sitecore-content-sdk/nextjs';
+import { ComponentProps } from '@/lib/component-props';
+
+type AlertBannerProps = ComponentProps &
+  AlertBannerParams &
+  AlertBannerFields &
+  AlertBannerData;
+
+type AlertBannerParams = {
+  params: {
+    mock_param?: string;
+  };
+};
+
+// Non-component data source fields
+// TODO_SCAFFOLD_BE: Populate if needed, remove if not
+type AlertBannerData = {
+  externalFields: {
+    mock_external_data: Field<string>;
+  };
+};
+
+type AlertBannerFields = {
+  fields: {
+    title: Field<string>; // Sitecore editable text field
+    description: Field<string>; // Sitecore editable text field
+    image?: ImageField; // Sitecore editable image field
+    link?: LinkField; // Sitecore editable link field
+  };
+};
 
 export const Default: React.FC<AlertBannerProps> = (props) => {
   const { fields } = props;
