+++ examples/kit-nextjs-article-starter/src/components/animated-section/AnimatedSection.dev.tsx
@@ -1,6 +1,25 @@
 import React, { useMemo } from 'react';
 import { useIntersectionObserver } from '@/hooks/use-intersection-observer';
-import { AnimatedSectionProps, StyleObject } from './animated-section.props';
+
+type AnimationType = 'slide' | 'rotate';
+type Direction = 'up' | 'down' | 'left' | 'right';
+
+interface AnimatedSectionProps {
+  children: React.ReactNode;
+  className?: string;
+  direction?: Direction;
+  delay?: number;
+  duration?: number;
+  animationType?: AnimationType;
+  endRotation?: number;
+  divWithImage?: React.RefObject<HTMLDivElement | null>;
+  reducedMotion?: boolean;
+  isPageEditing?: boolean;
+}
+
+interface StyleObject {
+  [key: string]: string | number;
+}
 
 export const Default: React.FC<AnimatedSectionProps> = React.memo(
   ({
