+++ examples/kit-nextjs-article-starter/src/components/article-header/ArticleHeader.tsx
@@ -5,9 +5,18 @@ import { useState, useEffect, useRef } from 'react';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { Facebook, Linkedin, Twitter, Link, Check, Mail } from 'lucide-react';
 import { cn } from '@/lib/utils';
-import { Text, DateField, useSitecore } from '@sitecore-content-sdk/nextjs';
+import {
+  Field,
+  Item,
+  ImageField,
+  LinkField,
+  Text,
+  DateField,
+  useSitecore,
+} from '@sitecore-content-sdk/nextjs';
+import { ComponentProps } from '@/lib/component-props';
+
 import { NoDataFallback } from '@/utils/NoDataFallback';
-import type { ArticleHeaderProps } from './article-header.props';
 import { Badge } from '@/components/ui/badge';
 import { Default as ImageWrapper } from '@/components/image/ImageWrapper.dev';
 import { Button } from '@/components/ui/button';
@@ -18,6 +27,61 @@ import { useI18n } from 'next-localization';
 import { dictionaryKeys } from '@/variables/dictionary';
 import { formatDateInUTC } from '@/utils/date-utils';
 import { Default as Icon } from '@/components/icon/Icon';
+
+interface ArticleHeaderParams {
+  [key: string]: any; // eslint-disable-line
+}
+
+type ReferenceField = {
+  id: string;
+  name: string;
+  url?: string;
+  displayName?: string;
+  fields?: {
+    [key: string]: Field | Item[] | ReferenceField | null;
+  };
+};
+
+type AuthorReferenceField = ReferenceField & {
+  fields: PersonItem;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+type AuthorItemFields = {
+  name: Field<string>;
+  jobTitle: Field<string>;
+};
+
+interface ArticleHeaderFields {
+  imageRequired?: { jsonValue: ImageField };
+  eyebrowOptional?: { jsonValue: Field<string> };
+}
+
+interface ArticleHeaderExternalFields {
+  pageHeaderTitle: { jsonValue: Field<string> };
+  pageReadTime?: { jsonValue: Field<string> };
+  pageDisplayDate?: { jsonValue: Field<string> };
+  pageAuthor?: { jsonValue: AuthorReferenceField };
+}
+
+interface ArticleHeaderProps extends ComponentProps {
+  params: ArticleHeaderParams;
+  fields: {
+    data: {
+      datasource: ArticleHeaderFields;
+      externalFields: ArticleHeaderExternalFields;
+    };
+  };
+}
+
+interface PersonItem {
+  personProfileImage?: ImageField;
+  personFirstName: Field<string>;
+  personLastName: Field<string>;
+  personJobTitle?: Field<string>;
+  personBio?: Field<string>;
+  personLinkedIn?: LinkField;
+}
 
 export const Default: React.FC<ArticleHeaderProps> = ({ fields }) => {
   const { imageRequired, eyebrowOptional } = fields.data?.datasource ?? {};
